Show loading state when uncommitting changes
Add withStackBusy wrapping around uncommit and uncommitChanges flows to
display a loading/busy state while uncommitting. This change ensures the
UI indicates activity when uncommitting from CommitView,
BranchCommitList, ChangedFilesContextMenu, and various drag/drop
handlers, and also centralizes update logic (selection updates and
uiState updates) to occur inside the busy wrapper.
